Refactor: rename deleted chats to expired messages
