+++ apps/desktop/desktop/src/main/java/bisq/desktop/main/content/chat/message_container/list/ChatMessagesListController.java
@@ -719,9 +719,10 @@ public class ChatMessagesListController implements Controller {
             addChatRulesWarningMessageListItemInPrivateChats(channel);
         }
 
-        boolean shouldShowDeletedMessagesIndicator = channel instanceof CommonPublicChatChannel || channel instanceof BisqEasyOfferbookChannel;
-        if (shouldShowDeletedMessagesIndicator) {
-            addDeletedChatsIndicator(channel);
+        boolean shouldShowExpiredMessagesIndicator = channel instanceof CommonPublicChatChannel
+                || channel instanceof BisqEasyOfferbookChannel;
+        if (shouldShowExpiredMessagesIndicator) {
+            addExpiredMessagesIndicator(channel);
         }
 
         maybeScrollDownOnNewItemAdded();
@@ -853,12 +854,12 @@ public class ChatMessagesListController implements Controller {
         }
     }
 
-    private <M extends ChatMessage, C extends ChatChannel<M>> void addDeletedChatsIndicator(C channel) {
+    private <M extends ChatMessage, C extends ChatChannel<M>> void addExpiredMessagesIndicator(C channel) {
         if (channel instanceof CommonPublicChatChannel commonPublicChatChannel) {
-            CommonPublicChatMessage commonPublicChatMessage = createDeletedChatsIndicator(commonPublicChatChannel);
+            CommonPublicChatMessage commonPublicChatMessage = createExpiredMessagesIndicator(commonPublicChatChannel);
             model.getChatMessages().add(createChatMessageListItem(commonPublicChatMessage, commonPublicChatChannel));
         } else if (channel instanceof BisqEasyOfferbookChannel bisqEasyOfferbookChannel) {
-            BisqEasyOfferbookMessage bisqEasyOfferbookMessage = createDeletedChatsIndicator(bisqEasyOfferbookChannel);
+            BisqEasyOfferbookMessage bisqEasyOfferbookMessage = createExpiredMessagesIndicator(bisqEasyOfferbookChannel);
             model.getChatMessages().add(createChatMessageListItem(bisqEasyOfferbookMessage, bisqEasyOfferbookChannel));
         }
     }
@@ -921,7 +922,7 @@ public class ChatMessagesListController implements Controller {
                 new HashSet<>());
     }
 
-    private CommonPublicChatMessage createDeletedChatsIndicator(CommonPublicChatChannel channel) {
+    private CommonPublicChatMessage createExpiredMessagesIndicator(CommonPublicChatChannel channel) {
         UserProfile senderUserProfile = userIdentityService.getSelectedUserIdentity().getUserProfile();
         long ttl_days = TimeUnit.MILLISECONDS.toDays(CommonPublicChatMessage.COMMON_PUBLIC_CHAT_MESSAGE_TTL);
         return new CommonPublicChatMessage(
@@ -929,15 +930,15 @@ public class ChatMessagesListController implements Controller {
                 channel.getChatChannelDomain(),
                 channel.getId(),
                 senderUserProfile.getId(),
-                Optional.of(Res.get("chat.public.deletedChatsIndicator.text", ttl_days)),
+                Optional.of(Res.get("chat.public.expiredMessagesIndicator.text", ttl_days)),
                 Optional.empty(),
                 System.currentTimeMillis() - TimeUnit.DAYS.toMillis(365), // Initialize date as 1y ago to always show as 1st message
                 false,
-                ChatMessageType.DELETED_CHATS_INDICATOR
+                ChatMessageType.EXPIRED_MESSAGES_INDICATOR
         );
     }
 
-    private BisqEasyOfferbookMessage createDeletedChatsIndicator(BisqEasyOfferbookChannel channel) {
+    private BisqEasyOfferbookMessage createExpiredMessagesIndicator(BisqEasyOfferbookChannel channel) {
         UserProfile senderUserProfile = userIdentityService.getSelectedUserIdentity().getUserProfile();
         long ttl_days = TimeUnit.MILLISECONDS.toDays(BisqEasyOfferbookMessage.BISQ_EASY_OFFERBOOK_MESSAGE_TTL);
         return new BisqEasyOfferbookMessage(
@@ -950,7 +951,7 @@ public class ChatMessagesListController implements Controller {
                 Optional.empty(),
                 0L,
                 false,
-                ChatMessageType.DELETED_CHATS_INDICATOR
+                ChatMessageType.EXPIRED_MESSAGES_INDICATOR
         );
     }
 
@@ -980,7 +981,7 @@ public class ChatMessagesListController implements Controller {
             if (chatMessage instanceof BisqEasyOfferbookMessage bisqEasyOfferbookMessage) {
                 switch (settingsService.getBisqEasyOfferbookMessageTypeFilter().get()) {
                     case OFFER -> isCorrectMessageType = bisqEasyOfferbookMessage.hasBisqEasyOffer()
-                            || bisqEasyOfferbookMessage.getChatMessageType() == ChatMessageType.DELETED_CHATS_INDICATOR;
+                            || bisqEasyOfferbookMessage.getChatMessageType() == ChatMessageType.EXPIRED_MESSAGES_INDICATOR;
                     case TEXT -> isCorrectMessageType = !bisqEasyOfferbookMessage.hasBisqEasyOffer();
                     case ALL -> isCorrectMessageType = true;
                     default ->
