+++ chat/src/main/java/bisq/chat/ChatMessageType.java
@@ -26,7 +26,7 @@ public enum ChatMessageType implements ProtoEnum {
     TAKE_BISQ_EASY_OFFER,
     PROTOCOL_LOG_MESSAGE,
     CHAT_RULES_WARNING,
-    DELETED_CHATS_INDICATOR;
+    EXPIRED_MESSAGES_INDICATOR;
 
     @Override
     public bisq.chat.protobuf.ChatMessageType toProtoEnum() {
