+++ i18n/src/main/resources/chat.properties
@@ -129,7 +129,7 @@ chat.private.chatRulesWarningMessage.onDismissChatRulesPopup.actionButtonText=Re
 # Public chats
 ######################################################
 
-chat.public.deletedChatsIndicator.text=Older messages were deleted. Messages in this channel expire after {0} days.
+chat.public.expiredMessagesIndicator.text=Older messages were deleted. Messages in this channel expire after {0} days.
 
 
 ######################################################
